`Proxy labels` substitute for labels
Proxy labels substitute for labels that aren't in the dataset. Proxy labels
are necessary when you can't directly measure what you want to predict. In
the video app, we can't directly measure whether or not a user will find a
video useful. It would be great if the dataset had a useful feature, and
users marked all the videos that they found useful, but because the dataset
doesn't, we'll need a proxy label that substitutes for usefulness.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>