Train your own model versus using a pre-trained model
It is always better to use a pre-trained model than training your own model
from the beginning. However, if the pre-trained model does not match the
features in your dataset then it is always better to write your own model
using the subsections of the pre-trained model to train your model.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>