Replace `Latex` "$" blocks with "```math```" in order to preserve formatting in GitHub
Currently, "Sundown" the official markdown parser for GitHub is unable to
parse a few latex syntax because of which the formatting gets disrupts in
some of the generated markdown files. So by properly replacing all of the
latex code blocks that are not parsed by the "Sundown" parser with the
"math" blocks will preserve the formatting of math equations.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>